% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esiD.r
\name{esiD}
\alias{esiD}
\title{Cohen's d Effect Size with Confidence Interval}
\usage{
esiD(m1, sd1, n1, m2, sd2, n2)
}
\arguments{
\item{m1}{Mean of group 1}

\item{sd1}{Standard deviation of group 1}

\item{n1}{Sample size of group 1}

\item{m2}{Mean of group 2}

\item{sd2}{Standard deviation of group 2}

\item{n2}{Sample size of group 2}
}
\value{
A vector with d, lower CI limit, and upper CI limit
}
\description{
Computes Cohen's d (Hedge's g) standardized mean difference with 95\%
confidence interval using weighted pooled standard deviation.
}
\details{
Uses weighted pooled SD to produce Hedge's g version of Cohen's d.
References: Hedge LV, Olkin I. Statistical methods for meta-analysis.
}
\examples{
esiD(0.7, 0.5, 317, 1.1, 0.8, 24)

}
