% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esiDse.r
\name{esiDse}
\alias{esiDse}
\title{Cohen's d Effect Size from Standard Error}
\usage{
esiDse(m1, se1, n1, m2, se2, n2)
}
\arguments{
\item{m1}{Mean of group 1}

\item{se1}{Standard error of group 1}

\item{n1}{Sample size of group 1}

\item{m2}{Mean of group 2}

\item{se2}{Standard error of group 2}

\item{n2}{Sample size of group 2}
}
\value{
The standardized mean difference (Cohen's d)
}
\description{
Computes Cohen's d (Hedge's g) standardized mean difference when
standard error is provided instead of standard deviation.
}
\examples{
esiDse(1.743, 0.062, 48, 1.787, 0.035, 144)

}
