% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esiP.r
\name{esiP}
\alias{esiP}
\title{Effect Size (h) from Two Proportions}
\usage{
esiP(p1, p2)
}
\arguments{
\item{p1}{First proportion (between 0 and 1)}

\item{p2}{Second proportion (between 0 and 1)}
}
\value{
The h effect size
}
\description{
Computes Cohen's h effect size from two proportions using
the arcsine transformation.
}
\details{
Computes h = |2\emph{arcsin(sqrt(p1)) - 2}arcsin(sqrt(p2))|
}
\examples{
esiP(0.6, 0.4)

}
