% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.r
\name{ess}
\alias{ess}
\title{Effective Sample Size for Clustered Data}
\usage{
ess(m, k, r)
}
\arguments{
\item{m}{Number of observations per cluster}

\item{k}{Number of clusters}

\item{r}{Intra-cluster correlation coefficient (ICC)}
}
\value{
A vector with two elements: design effect and effective sample size
}
\description{
Calculates the design effect and effective sample size for clustered data
(e.g., repeated measurements on subjects).
}
\details{
Design effect = (m-1)*r + 1
When dividing total sample size by the design effect, you get the effective
sample size that would be needed for simple random sampling to achieve the
same power and precision as the clustered design.

References: van Breukelen & Candel (2012). Journal of Clinical Epidemiology.
}
\examples{
ess(25, 4, 0.017)  # 25 obs per cluster, 4 clusters, ICC = 0.017

}
