% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_discrete.r
\name{hist_discrete}
\alias{hist_discrete}
\title{Histogram for Discrete (Integer) Data}
\usage{
hist_discrete(data, variable = NULL)
}
\arguments{
\item{data}{A data frame or numeric vector}

\item{variable}{The variable name to plot (if data is a data frame)}
}
\description{
Creates a histogram with breaks aligned to integer values, useful for
visualizing discrete or count data.
}
\details{
Automatically sets breaks at integer boundaries for cleaner display of discrete data.
Can work with either a data frame and variable name or a direct vector.
}
\examples{
\dontrun{
hist_discrete(mtcars, cyl)
hist_discrete(c(1, 1, 2, 2, 3, 3, 3, 4, 4, 5))
}

}
