% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.r
\name{include}
\alias{include}
\title{Source File and Save Output}
\usage{
include(filename, output_suffix = "txt")
}
\arguments{
\item{filename}{Path to the R script file to source}

\item{output_suffix}{File extension for the output file (default "txt")}
}
\description{
Sources an R script file and saves both the commands and their output to a text file.
}
\details{
Executes an R script and captures all commands and their output to a file
with the same name as the input file but with the specified suffix.
Useful for creating reproducible analysis logs.
}
\examples{
\dontrun{
include("myscript.R")
# Creates myscript.txt with commands and output
}

}
