% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neff.r
\name{neff}
\alias{neff}
\title{Effective Sample Size (Harmonic Mean)}
\usage{
neff(n1, n2)
}
\arguments{
\item{n1}{Sample size of group 1}

\item{n2}{Sample size of group 2}
}
\value{
The harmonic mean of n1 and n2
}
\description{
Calculates the effective per-group sample size as the harmonic mean of two group sizes.
}
\details{
neff = 2 / (1/n1 + 1/n2)
}
\examples{
neff(20, 30)

}
