% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooledSD.r
\name{pooledSD}
\alias{pooledSD}
\title{Pooled Standard Deviation}
\usage{
pooledSD(sd1, sd2, n1, n2)
}
\arguments{
\item{sd1}{Standard deviation of group 1}

\item{sd2}{Standard deviation of group 2}

\item{n1}{Sample size of group 1}

\item{n2}{Sample size of group 2}
}
\value{
The pooled standard deviation
}
\description{
Computes the pooled standard deviation for two independent groups.
}
\details{
pooled_sd = sqrt(((n1-1)*sd1^2 + (n2-1)*sd2^2) / (n1 + n2 - 2))
}
\examples{
pooledSD(0.43, 0.442, 48, 144)

}
