% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvf.r
\name{pvf}
\alias{pvf}
\title{Format P-values for Academic Reporting}
\usage{
pvf(p)
}
\arguments{
\item{p}{A p-value (between 0 and 1)}
}
\value{
A character string with formatted p-value
}
\description{
Formats p-values according to academic publication guidelines.
}
\details{
Follows standard academic formatting:
\itemize{
\item P > .999: "> .99"
\item P >= .01: 2 decimal places (or 3 if between .045 and .049)
\item .001 <= P < .01: 3 decimal places
\item P < .001: "< .001"
}
}
\examples{
pvf(0.046)
pvf(0.001)
pvf(0.00001)

}
