% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showme.r
\name{showme}
\alias{showme}
\title{Display File Contents}
\usage{
showme(x)
}
\arguments{
\item{x}{Path to the text file}
}
\description{
Reads and displays the contents of a text file to the console.
}
\examples{
\dontrun{
showme("/path/to/file.txt")
}

}
