% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showmeSectionLines.r
\name{showmeSectionLines}
\alias{showmeSectionLines}
\title{Display Specific Number of File Lines}
\usage{
showmeSectionLines(x, search_string = NULL, lines = 999)
}
\arguments{
\item{x}{Path to the text file}

\item{search_string}{Optional string to search for as a starting point}

\item{lines}{Number of lines to display (default 999)}
}
\description{
Displays a specified number of lines from a text file starting from an optional search string.
}
\details{
If search_string is provided, starts display from that line; otherwise starts from line 1.
}
\examples{
\dontrun{
showmeSectionLines("/path/to/file.txt", "START", lines = 10)
}

}
