% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_https.r
\name{source_https}
\alias{source_https}
\title{Source R Scripts from HTTPS URLs}
\usage{
source_https(url, ...)
}
\arguments{
\item{url}{The HTTPS URL of the R script to source}

\item{...}{Additional URLs to source}
}
\description{
Sources R scripts hosted on GitHub or other HTTPS URLs directly into the R environment.
}
\details{
Uses RCurl package to download and parse R scripts from HTTPS sources.
Useful for sourcing functions from GitHub repositories.
}
\examples{
\dontrun{
source_https("https://raw.github.com/user/repo/branch/file.R")
}

}
