% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svalues.r
\name{svalues}
\alias{svalues}
\title{Extract SVALUES from Mplus Output File}
\usage{
svalues(filename)
}
\arguments{
\item{filename}{Path to the Mplus output file}
}
\value{
A character vector containing the model parameters with final estimate values
}
\description{
Extracts the SVALUES section from an Mplus output file to retrieve starting values
for model parameters.
}
\details{
Parses Mplus output to extract the section containing model commands with final
estimates that can be used as starting values for subsequent analyses.
}
\examples{
\dontrun{
svalues_import <- svalues("mymodel.out")
}

}
