% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttesti.r
\name{ttesti}
\alias{ttesti}
\title{Two-Tailed P-value from t-statistic}
\usage{
ttesti(x, df)
}
\arguments{
\item{x}{The t-statistic value}

\item{df}{Degrees of freedom}
}
\value{
The two-tailed p-value
}
\description{
Calculates two-tailed p-value from a t-statistic and degrees of freedom.
}
\examples{
ttesti(2.5, 28)

}
