% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varnames.r
\name{varnames}
\alias{varnames}
\title{Print Variable Names from Data Frame}
\usage{
varnames(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
Prints variable names separated by commas
}
\description{
Prints the names of a data frame as a comma-separated string on a single line.
}
\details{
Useful for quickly viewing all variable names in a data frame.
Works with both native pipe and standard function calls.
}
\examples{
iris |> varnames()
varnames(mtcars)

}
