% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vDd.r
\name{vDd}
\alias{vDd}
\title{Verbal Description of Cohen's d Effect Size}
\usage{
vDd(d)
}
\arguments{
\item{d}{Cohen's d effect size}
}
\value{
A character string describing the effect size magnitude
}
\description{
Provides a qualitative descriptor for a given Cohen's d effect size.
}
\details{
Converts numerical d values to verbal descriptors:
\itemize{
\item d < .2: "trivial"
\item .2 <= d < .25: "small"
\item .25 <= d < .45: "small-to-medium"
\item .45 <= d < .55: "about a medium"
\item .55 <= d < .75: "medium-to-large"
\item .75 <= d < .85: "about a large"
\item d >= .85: "large"
}
}
\examples{
vDd(0.5)   # "about a medium"
vDd(0.8)   # "about a large"
vDd(1.0)   # "large"

}
