% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vDr.r
\name{vDr}
\alias{vDr}
\title{Verbal Description of Correlation Coefficient}
\usage{
vDr(r)
}
\arguments{
\item{r}{Pearson correlation coefficient (between -1 and 1)}
}
\value{
A character string describing the correlation magnitude
}
\description{
Provides a qualitative descriptor for a given correlation coefficient.
}
\details{
Converts |r| to verbal descriptors:
\itemize{
\item |r| < .1: "trivial"
\item .1 <= |r| < .12: "small"
\item .12 <= |r| < .28: "small-to-medium"
\item .28 <= |r| < .32: "about a medium"
\item .32 <= |r| < .48: "medium-to-large"
\item .48 <= |r| < .52: "about a large"
\item |r| >= .52: "large"
}
}
\examples{
vDr(0.3)   # "about a medium"
vDr(0.5)   # "about a large"
vDr(0.7)   # "large"

}
