% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztesti.r
\name{ztesti}
\alias{ztesti}
\title{Two-Tailed P-value from z-statistic}
\usage{
ztesti(x)
}
\arguments{
\item{x}{The z-statistic value}
}
\value{
The two-tailed p-value
}
\description{
Calculates two-tailed p-value from a z-statistic.
}
\examples{
ztesti(2.5)

}
