% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z1.r
\name{z1}
\alias{z1}
\title{Z1 Score Transformation}
\usage{
z1(
  data,
  vars,
  method = "default",
  lambda = 2,
  logit = FALSE,
  z = FALSE,
  altmax = NULL,
  silent = TRUE
)
}
\arguments{
\item{data}{A data frame containing the variables to transform}

\item{vars}{Character vector of variable names to transform}

\item{method}{One of "default", "softmax", or "altmethod". Default is "default".}

\item{lambda}{For softmax method: number of standard deviations over which the
transformed variable is expected to be linear. Default is 2.}

\item{logit}{If TRUE, apply logit transformation to the (0,1) scaled variable}

\item{z}{If TRUE, apply inverse normal transformation (z-score) to the scaled variable}

\item{altmax}{Alternative maximum value for scaling (overrides observed maximum)}

\item{silent}{If FALSE, print back-transformation formulas}
}
\value{
A data frame with the original variables and additional \verb{_z1} variables
containing the transformed values. Attributes contain the scaling parameters.
}
\description{
Transforms variables to a (0,1) scale with optional further transformations.
}
\details{
The default method scales variables using:
\code{(x - min + (5/(8*1000))*sd) / (max - min + (10/(8*1000))*sd)}

The softmax method uses a sigmoid transformation based on the mean and standard deviation.

The altmethod uses a more complex approach involving regression on collapsed data.
}
\examples{
\dontrun{
data <- data.frame(x = 1:10, y = seq(0, 100, length.out = 10))
result <- z1(data, vars = c("x", "y"))
head(result)
}

}
